+++ src/main/java/commoble/bagofyurting/ObjectNames.java
@@ -6,5 +6,7 @@ public class ObjectNames
 	public static final String SHAPED_UPGRADE_RECIPE = "shaped_bag_of_yurting_upgrade";
 	public static final String SHAPELESS_UPGRADE_RECIPE = "shapeless_bag_of_yurting_upgrade";
 	public static final String UPGRADE_RECIPE = "bag_of_yurting_upgrade";
+	public static final String RADIUS = "radius";
+	public static final String YURTDATA = "yurtdata";
 
 }