+++ src/main/java/commoble/bagofyurting/util/ConfigHelper.java
@@ -24,51 +24,37 @@ import java.util.function.Function;
 
 import org.jetbrains.annotations.Nullable;
 
-import net.minecraftforge.common.ForgeConfigSpec;
-import net.minecraftforge.fml.ModLoadingContext;
-import net.minecraftforge.fml.config.ModConfig;
-
-
+import net.neoforged.fml.ModList;
+import net.neoforged.fml.config.ModConfig;
+import net.neoforged.neoforge.common.ModConfigSpec;
 
 public class ConfigHelper
 {
 	public static <T> T register(
+		final String modid,
 		final ModConfig.Type configType,
-		final Function<ForgeConfigSpec.Builder, T> configFactory)
+		final Function<ModConfigSpec.Builder, T> configFactory)
 	{
-		return register(configType, configFactory, null);
+		return register(modid, configType, configFactory, null);
 	}
-	
-	/**
-	 * Register a config using a custom filename.
-	 * @param <T> Your config class
-	 * @param configType Forge config type:
-	 * <ul>
-	 * <li>SERVER configs are defined by the server and synced to clients; individual configs are generated per-save.
-	 * <li>COMMON configs are definable by both server and clients and not synced (they may have different values)
-	 * <li>CLIENT configs are defined by clients and not used on the server
-	 * </ul>
-	 * @param configFactory A constructor or factory for your config class
-	 * @param configName Name of your config file. Supports subfolders, e.g. "yourmod/yourconfig".
-	 * @return An instance of your config class
-	 */
 	public static <T> T register(
+		final String modid,
 		final ModConfig.Type configType,
-		final Function<ForgeConfigSpec.Builder, T> configFactory,
+		final Function<ModConfigSpec.Builder, T> configFactory,
 		final @Nullable String configName)
 	{
-		final ModLoadingContext modContext = ModLoadingContext.get();
-		final org.apache.commons.lang3.tuple.Pair<T, ForgeConfigSpec> entry = new ForgeConfigSpec.Builder()
+		final var mod = ModList.get().getModContainerById(modid).get();
+		final org.apache.commons.lang3.tuple.Pair<T, ModConfigSpec> entry = new ModConfigSpec.Builder()
 			.configure(configFactory);
 		final T config = entry.getLeft();
-		final ForgeConfigSpec spec = entry.getRight();
+		final ModConfigSpec spec = entry.getRight();
 		if (configName == null)
 		{
-			modContext.registerConfig(configType,spec);
+			mod.registerConfig(configType,spec);
 		}
 		else
 		{
-			modContext.registerConfig(configType, spec, configName + ".toml");
+			mod.registerConfig(configType, spec, configName + ".toml");
 		}
 		
 		return config;